ioPath - make OSS black version match FB black version (#5)
Summary:
Pull Request resolved: https://github.com/fairinternal/iopath/pull/5

While `black --check .` is passing locally in fbcode, the same command is failing in OSS. This is because the `black` versions are different.

fbcode uses `black==20.8b1`. This diff locks `black` at `20.8b1` in the CircleCI config. You can see that many other FB OSS repos also lock `black` at this version. The only repo that uses the custom version of `black` that ioPath is currently set to is AIRStore.

Reviewed By: sujitoc

Differential Revision: D27337884

fbshipit-source-id: aa990de394ed1ee26d51106404cc66dfe54baa33