Support for custom setup of a path handler.
Summary:
Support for custom setup of a path handler.
This allows common code with oss.

This change also required reversing of the override default in register_handler as otherwise it would lead to errors in existing code.

Please note that `custom/setup_handlers.py` will not be exported to OSS. It is only meant to register Manifold and Everstore handlers in FB env.

Reviewed By: hudeven

Differential Revision: D27960781

fbshipit-source-id: ae0b1912d8b671380eb04cbdb0dbaa4da3c8503f