Allow copy across storages
Summary:
Currently, we cannot copy directly from storage to another.
The only use case where we would want this anyway is between local and manifold (both ways), that's true. But it forces us to use copy_from_local or get_local_path instead, or open('rb') with open('wb') and transferring the contents. It would be nicer if the user just doesn't have to worry about which method to call.

Reviewed By: sujitoc

Differential Revision: D30101469

fbshipit-source-id: 4d3b817be18c07bc8da5e0ec29d7ff24d08a6431