Lint and black-format fair_infra/data/iopath
Summary: Currently, the `fbcode/fair_infra/` directory is opted-out of Python-black autoformatting. I care about some files in that directory - I want them to look nice - so I'd like to remove the opt-out. This diff pre-formats some code in that directory so that when I eventually remove the opt-out, I won't get a bunch of generated changes.

Reviewed By: zertosh

Differential Revision: D31952917

fbshipit-source-id: 8d341d98a1c07e20968c694c2b643270bcff992c