Introduce a new context manager interface to be used by opent()
Summary:
We would like to enforce static type checks for `PathHandler`
implementations that implement the `opent()` method. Currently, `opent()` has
`Iterable[Any]` as its return type annotation which makes it difficult for
Pyre to carry out static type checks of the method since the type is resolved
at runtime. We can change the method's signature to return a context manager
instance instead which we can then use to enforce the type check.

This diff introduces the `TabularIO` interface which behaves
as a context manager and can be returned by `opent()` implementations.
We will use this in all implementations of `opent()` in following diffs.

Reviewed By: mackorone

Differential Revision: D33820881

fbshipit-source-id: a500035089a335c22916a6424be84a8afa7ab89b