iopath overload
Summary:
Add the overloads that are missing for "open". We can (and should) use the `Literal` type to tell pyre whether the return type of `open` is a string or a bytestring.

See [PEP 586](https://peps.python.org/pep-0586/#shortening-unions-of-literals) which actually has basically exactly this overload defined as a canonical usage of the `Literal` type.

Reviewed By: luccabb

Differential Revision: D40639404

fbshipit-source-id: 14f9fe328483b57393e6584809bd7c1f2bc74148