Fix race condition in EventLogger
Summary:
Before this commit, the code allowed for a race condition when on
thread1, `del self._evt` is executed but `self._evt = SimpleEventRecord()` is
not yet, and thread2 tries to access `self._evt`. This leads to errors reported
in e.g. https://fb.workplace.com/groups/ManifoldUsers/permalink/1960108557737182/

This commit fixes the race condition: `del` instruction was reduntant, because
both `del self._evt` and `self._evt = SimpleEventRecord()` decrement the
reference count on the old object.

Differential Revision: D59416308

fbshipit-source-id: 1d327224243dea4f7ab5c44f6a9f6145a25107ed