add an open api for write
Summary:
we need to allow client writing bundles for the image sequence cutting workflow. This diff adds a new api specifically for write. All writing codes with be in a separate path so it will minimize the impact to the existing read path.

I can't use the `_open` because it's integrated with fb scuba logging which can't work in the OSS env. removing scuba will require changes to the NativePathHandler which seems to be used somewhere.

The consumer of this api will be very limited to only the image sequence cutting workflow in the xlformers. We can easily deprecate this code when we have a better solution in the future.

Reviewed By: ameya0111

Differential Revision: D74012420

fbshipit-source-id: f8ff9114d6132c75cec7ac06e1d21300be5357be